Refactoring module naming
