+++ src/components/favorites-screen/favorites-screen.jsx
