+++ src/components/offer-screen/offer-screen.jsx
