+++ src/components/place-card-cities/place-card-cities.jsx
@@ -1,5 +1,5 @@
 import React, {PureComponent} from "react";
-import PlaceCardScreen from "../place-card/place-card-screen.jsx";
+import PlaceCardScreen from "../place-card/place-card.jsx";
 
 class PlaceCardCities extends PureComponent {
   constructor(props) {
