+++ src/components/place-card-neighbour/place-card-neighbour.jsx
@@ -1,5 +1,5 @@
 import React, {PureComponent} from "react";
-import PlaceCardScreen from "../place-card/place-card-screen.jsx";
+import PlaceCardScreen from "../place-card/place-card.jsx";
 
 class PlaceCardNeighbour extends PureComponent {
   constructor(props) {
