+++ src/components/place-card/place-card.jsx
